code/requirements.txt
numpy=1.24.2
pycryptodome=3.19.0
watchdog=3.0.0
obspy=1.4.0
